added changes

+++ backend/src/services/imageService.js
@@ -1,6 +1,7 @@
 const { StatusCodes } = require('http-status-codes');
 const imagesModel = require('../models/imagesModel');
 const { formatErrorMsg } = require('../utils/CommonFuncs');
+const bcrypt = require('bcryptjs');
 
 const getList = async (params, user) => {
   const result = {
@@ -11,6 +12,7 @@ const getList = async (params, user) => {
     const skip = Number(params.skip) || 0;
     const searchString = params.searchString || '';
     const data = await imagesModel.fetchAll(skip, searchString, user.id);
+    console.log(data);
     result.data = data;
   } catch (e) {
     console.log(e);
@@ -96,21 +98,31 @@ const updateById = async (params, data) => {
   return result;
 };
 
-const deleteById = async (params) => {
+const deleteById = async (req) => {
   const result = {
     error: false,
     data: {},
   };
-  const id = Number(params.id) || 0;
+  const id = Number(req.params.id) || 0;
 
   try {
-    const qData = await imagesModel.deleteById(id);
-    if (qData) {
-      result.data = qData;
+    const isPasswordMatch = await bcrypt.compare(
+      req.body.confirmedPassword || '',
+      req.user.password
+    );
+    if (isPasswordMatch) {
+      const qData = await imagesModel.deleteById(id);
+      if (qData) {
+        result.data = qData;
+      } else {
+        result.error = true;
+        result.status = StatusCodes.NOT_FOUND;
+        result.message = 'No Image Found.';
+      }
     } else {
       result.error = true;
-      result.status = StatusCodes.NOT_FOUND;
-      result.message = 'No Todo Found.';
+      result.status = StatusCodes.BAD_REQUEST;
+      result.message = 'The provided value for the password is invalid. You cannot delete image.';
     }
   } catch (e) {
     result.error = true;
